Stop merging/unmerging profiles randomly, simply encourage profiles which don't derive from base or user and have an :android-config in your profiles for setting the sdk and any injections or dependencies which you want to include in all your android projects.
This also allows the :android :build-type :debug or :release option,
rather than examining which profiles were included, this is far more
robust and allows you to have multiple release and debug
profiles. Generally this is more predictable and in line with the way
that leiningen profiles are supposed to be used.

I have left the doall and release tasks for now for back compatibility
and quick-start, but more and more they should be replaced my profile
inheritance and simple aliases in the future.